More clean-up.
