--- src/main/java/lykrast/voyage/biomebuilder/BiomeTemplate.java
@@ -1,14 +1,0 @@
-package lykrast.voyage.biomebuilder;
-
-public final class BiomeTemplate {
-  private final TerraformBiomeBuilder builder;
-
-  public BiomeTemplate(TerraformBiomeBuilder builder) {
-    this.builder = builder;
-    builder.markTemplate();
-  }
-
-  public TerraformBiomeBuilder builder() {
-    return new TerraformBiomeBuilder(this.builder);
-  }
-}
